[bug][feature]. remove file instead of directory. add author_name, author_email, description into tempalte. add pip.conf moving. remove all layouts readme. add empty readme

+++ tamplar/__internal/utils.py
@@ -8,24 +8,13 @@ input_ = input
 
 def mv(src, dst, symlinks=False, ignore=None):
     for item in os.listdir(src):
-        s = os.path.join(src, item)
-        d = os.path.join(dst, item)
+        s = os.path.abspath(os.path.join(src, item))
+        d = os.path.abspath(os.path.join(dst, item))
         if os.path.isdir(s):
             shutil.copytree(s, d, symlinks, ignore)
         else:
             shutil.copy2(s, d)
     shutil.rmtree(src)
-
-
-def name_validator(project_name):
-    name_symbols = set(project_name.lower())
-    all_symbols = string.ascii_lowercase + string.digits + ' -_'
-    intersection = name_symbols.difference(all_symbols)
-    assert len(intersection) == 0, 'invalid project name'
-
-
-def package_name(name):
-    return name.replace('-', '_').replace(' ', '_')
 
 
 def empty_folder(objs):
@@ -43,11 +32,14 @@ def clean_directory(path, agree=None):
         return True
     clean = agree
     if agree is None:
-        clean = input_("Directory is not empty. All files will be remove [Y/n]:")
+        clean = input_('directory is not empty. All files will be remove [Y/n]: ')
     clean = clean.lower()
     assert clean in ['y', 'n', '']
     if clean == 'n':
         return False
     for obj in objs:
-        shutil.rmtree(path+obj)
+        if os.path.isdir(path+obj):
+            shutil.rmtree(path+obj)
+            continue
+        os.remove(path+obj)
     return True
